feat: initialize demo project with Vite and WASM support
