--- demo/package.json
@@ -1,18 +1,0 @@
-{
-    "scripts": {
-        "serve": "webpack-dev-server",
-        "build": "rimraf dist pkg && webpack",
-        "start": "rimraf dist pkg && webpack-dev-server --hot",
-        "test": "cargo test && wasm-pack test --headless"
-    },
-    "devDependencies": {
-        "@wasm-tool/wasm-pack-plugin": "^1.1.0",
-        "copy-webpack-plugin": "^5.0.3",
-        "http-server": "^14.1.1",
-        "rimraf": "^3.0.0",
-        "rust-wasm-loader": "^0.2.0",
-        "webpack": "^5.94.0",
-        "webpack-cli": "^3.1.2",
-        "webpack-dev-server": "^3.1.10"
-    }
-}
