+++ packages/kos-web/demo/index.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>KOS-Rust example</title>
+  </head>
+  <body>
+    <script type="module" src="/src/main.ts"></script>
+
+    <button id="btnCreateMnemonic">Create Mnemonic</button>
+    <button id="btnLoadMnemonic">Load Mnemonic</button>
+    <button id="btnSendKLV">Send KLV</button>
+    <button id="btnWMFlow">Wallet Manager Flow</button>
+    <hr />
+    <img
+      id="qrCode"
+      alt="Encrypted Mnemonic"
+      style="width: 50%; display: none"
+    />
+    <hr />
+    <p id="demo"></p>
+
+    <input type="file" id="fileInput" accept="image/*" style="display: none" />
+  </body>
+</html>
