+++ packages/kos-web/demo/package.json
@@ -1,0 +1,20 @@
+{
+  "name": "demo",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview"
+  },
+  "devDependencies": {
+    "typescript": "~5.7.2",
+    "vite": "^6.2.0"
+  },
+  "dependencies": {
+    "jsqr": "1.4.0",
+    "kos": "file:./kos",
+    "vite-plugin-wasm": "^3.4.1"
+  }
+}
