+++ packages/kos-web/demo/vite.config.ts
@@ -1,0 +1,6 @@
+import { defineConfig } from "vite";
+import wasm from "vite-plugin-wasm";
+
+export default defineConfig({
+  plugins: [wasm()],
+});
